requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2024.2.2=py38h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
ctcdecode=1.0.3=pypi_0
cuda=11.6.1=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=12.3.101=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.3.101=0
cuda-nsight-compute=12.3.2=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=12.3.101=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=12.3.101=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=12.3.101=0
cuda-runtime=11.6.1=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=12.3.101=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-visual-tools=11.6.1=0
ffmpeg=4.3=hf484d3e_0
freetype=2.12.1=h4a9f257_0
gds-tools=1.8.1.2=0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.28.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.62.0=pypi_0
idna=3.4=py38h06a4308_0
importlib-metadata=7.0.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.9.2.110=h5e84587_0
libcublas-dev=11.9.2.110=h5c901ab_0
libcufft=10.7.1.112=hf425ae0_0
libcufft-dev=10.7.1.112=ha5ce4c0_0
libcufile=1.8.1.2=0
libcufile-dev=1.8.1.2=0
libcurand=10.3.4.107=0
libcurand-dev=10.3.4.107=0
libcusolver=11.3.4.124=h33c3c4e_0
libcusparse=11.7.2.124=h7538f96_0
libcusparse-dev=11.7.2.124=hbbe9722_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.6.3.124=hd2722f0_0
libnpp-dev=11.6.3.124=h3c42840_0
libnvjpeg=11.6.2.124=hd473ad6_0
libnvjpeg-dev=11.6.2.124=hb5906b9_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.2=pypi_0
markupsafe=2.1.5=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
nsight-compute=2023.3.1.1=0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
pillow=10.2.0=py38h5eee18b_0
pip=23.3.1=py38h06a4308_0
protobuf=4.25.3=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
python=3.8.18=h955ad1f_0
pytorch=1.13.1=py3.8_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=py38h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.31.0=py38h06a4308_1
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.13.1=py38_cu116
torchvision=0.14.1=py38_cu116
tqdm=4.66.2=pypi_0
typing_extensions=4.9.0=py38h06a4308_1
urllib3=2.1.0=py38h06a4308_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.5=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
